pyproject.toml
[project]
name = "subscribe-manager"
version = "0.3.2"
description = "订阅代理工具，统一管理订阅，定时下载并存储订阅信息，自动进行订阅转换，并提供代理服务"
authors = [{name = "modestme", email = "lowkeymodestme@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent"
]
license = { text = "MIT" }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "subscribe_manager", from = "src" },
]
[tool.poetry.scripts]
sbscmgr="subscribe_manager.main:main"

[[tool.poetry.source]]
name = "pypi_mirror"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[tool.poetry.dependencies]
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
pyyaml = "^6.0.2"
apscheduler = "^3.10.4"
pydantic-settings = "^2.6.0"
aiohttp = "^3.10.10"
nest-asyncio = "^1.6.0"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
mypy = "^1.11.2"
types-pyyaml = "^6.0.12.20250516"

[tool.ruff]
target-version = "py311"
exclude = [".venv", "tests", "build"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "N"]
pydocstyle.convention = "google"

[tool.mypy]
ignore_missing_imports = true
strict = true
disallow_untyped_defs = true
check_untyped_defs = true
exclude = "^(tests/|build/|dist/)"